% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jNumberTheory.R
\name{continuedFraction}
\alias{continuedFraction}
\title{Give the continued fraction representation}
\usage{
continuedFraction(x, nmax = 10)
}
\arguments{
\item{x}{Real}

\item{nmax}{Integer. Maximum number of terms to return}
}
\description{
Give the continued fraction representation
}
