% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\docType{class}
\name{Cylinder-class}
\alias{Cylinder-class}
\title{An S4 class representing a cylinder and how a lattice should be displayed on it.
Node zero of the lattice has coordinates (0,0).  We display a window of size -1/2<x<1/2, 0<z<L, and by default node zero is at the centre of
this window, but this can be changed with the origin slot.}
\description{
An S4 class representing a cylinder and how a lattice should be displayed on it.
Node zero of the lattice has coordinates (0,0).  We display a window of size -1/2<x<1/2, 0<z<L, and by default node zero is at the centre of
this window, but this can be changed with the origin slot.
}
\section{Slots}{

\describe{
\item{\code{Rho}}{Radius of the cylinder unused}

\item{\code{L}}{display height}

\item{\code{origin}}{eg (-1/2,-L/2) will place node 0 at the bottom left of the display window}

\item{\code{nRepeats}}{number of time to repeat periodically left and right NOT RELIABLE}
}}

