% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyllotaxisGeneratingIntervals.R
\name{cylindricalX}
\alias{cylindricalX}
\title{Cylindrical coordinate
 @param m Integer.}
\usage{
cylindricalX(m, d, scale = "unity")
}
\arguments{
\item{d}{Divergence of the lattice. 0<d<1 if scale=unity}

\item{scale}{twomn or unity}
}
\description{
Cylindrical coordinate
 @param m Integer.
}
