% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyllotaxisGeneratingIntervals.R
\name{DeltaForLHSd}
\alias{DeltaForLHSd}
\title{Return the Delta for which the generating interval (m,n,Delta) is in 0<d<1/2}
\usage{
DeltaForLHSd(m, n)
}
\arguments{
\item{m, n}{Integer}
}
\description{
Return the Delta for which the generating interval (m,n,Delta) is in 0<d<1/2
}
