% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyllotaxisGeneratingIntervals.R
\name{Deltamnd}
\alias{Deltamnd}
\title{The sign of the area of the generating parallelogram}
\usage{
Deltamnd(m, n, d)
}
\arguments{
\item{m, n}{Integer. Orders of the generating parastichy}

\item{d}{Divergence}
}
\description{
Interpreting this as an area assumes that the rise is equal to one
}
