% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyllotaxisGeneratingIntervals.R
\name{dGeneratingOpposedIntervalLHS}
\alias{dGeneratingOpposedIntervalLHS}
\title{Return the interval for which m,n is generating and opposed in 0<d<1/2}
\usage{
dGeneratingOpposedIntervalLHS(m, n)
}
\arguments{
\item{m, n}{Integer}
}
\description{
Return the interval for which m,n is generating and opposed in 0<d<1/2
}
