% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{dhvanIterson}
\alias{dhvanIterson}
\title{The divergence-rise curve in a van Iterson diagram for m,n}
\usage{
dhvanIterson(m, n, ontype, lprtype)
}
\arguments{
\item{m, n}{Integer. Orders of a generating parastichy pair}

\item{d}{Divergence of the lattice. 0<d<1.}
}
\description{
The divergence-rise curve in a van Iterson diagram for m,n
}
