% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{dhvanItersonPseudoGenerating}
\alias{dhvanItersonPseudoGenerating}
\title{The point in d,h space where a touching circles lattice becomes pseudogenerating}
\usage{
dhvanItersonPseudoGenerating(m)
}
\arguments{
\item{m}{Integer}
}
\description{
The point in d,h space where a touching circles lattice becomes pseudogenerating
}
