% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyllotaxisGeneratingIntervals.R
\name{divergenceInterval}
\alias{divergenceInterval}
\title{The range of possible divergences for a parastichy pair of a given type}
\usage{
divergenceInterval(m, n, Delta, scale = "unity", type = "generating")
}
\arguments{
\item{m, n}{Integer parastichy orders}

\item{Delta}{+/- 1}

\item{scale}{If "unity", the divergence (between 0 and 1); if "twomn" the divergence multiplied by 2*m*n}

\item{type}{One of "generating", "pseudogenerating", "opposed", or "generatingopposed"}
}
\description{
If a parastichy pair is generating or opposed there is a constraint on the ange of possible divergences
which this function returns
}
