% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\name{getRho}
\alias{getRho}
\title{The unscaled radius of the cylinder}
\usage{
getRho(object, Jugacy = 1)
}
\arguments{
\item{Cylinder}{an object inheriting from class Cylinder}
}
\description{
The unscaled radius of the cylinder
}
