% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\name{getXlim}
\alias{getXlim}
\title{Range of horizontal coordinate in the plotted cylindrical lattice}
\usage{
getXlim(Cylinder)
}
\arguments{
\item{Cylinder}{an object inheriting from class Cylinder}
}
\description{
Range of horizontal coordinate in the plotted cylindrical lattice
}
