% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\name{getYlim}
\alias{getYlim}
\title{Range of vertical coordinate in the plotted cylindrical lattice}
\usage{
getYlim(Cylinder)
}
\arguments{
\item{Cylinder}{an object inheriting from class Cylinder}
}
\description{
Range of vertical coordinate in the plotted cylindrical lattice
}
