% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{hvanIterson}
\alias{hvanIterson}
\title{The rise for a given divergence in a van Iterson curve for m,n}
\usage{
hvanIterson(m, n, d)
}
\arguments{
\item{m, n}{Integer. Orders of a generating parastichy pair}

\item{d}{Divergence of the lattice. 0<d<1.}
}
\description{
The rise for a given divergence in a van Iterson curve for m,n
}
