% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{intervalDelete}
\alias{intervalDelete}
\title{The difference of two intervals}
\usage{
intervalDelete(a, b)
}
\arguments{
\item{a, b}{Vectors of length two representing real intervals}
}
\description{
Throws an error if the resulting interval is disjoint
}
