% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyllotaxisGeneratingIntervals.R
\name{isOpposed}
\alias{isOpposed}
\title{Check if a parastichy pair is opposed}
\usage{
isOpposed(m, n, d, scale = "twomn")
}
\arguments{
\item{d}{Divergence of the lattice. 0<d<1 if scale=unity}

\item{scale}{twomn or unity}
}
\description{
Used by test suite
 @param m,n Integer. Orders of a generating parastichy pair
}
