% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JeanComputation.R
\name{JeanComputation}
\alias{JeanComputation}
\title{Perform the computation of Jean(1994) p40}
\usage{
JeanComputation(divergence, kmax = 21, doOpposed = TRUE)
}
\description{
Perform the computation of Jean(1994) p40
}
