% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\name{latticePoint}
\alias{latticePoint}
\title{Coordinates of a point in the cylindrical lattice within the display window}
\usage{
latticePoint(PG, n)
}
\arguments{
\item{PG}{A lattice object}

\item{n}{Sequence number of the point to return}
}
\description{
Coordinates of a point in the cylindrical lattice within the display window
}
