% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\name{newPhyllotaxisGenetic}
\alias{newPhyllotaxisGenetic}
\title{Create a new PhyllotaxisGenetic object}
\usage{
newPhyllotaxisGenetic(Rise = 1, Divergence, Rho = 1/(2 * pi), Jugacy = 1,
  L, Cylinder, ...)
}
\arguments{
\item{Rise}{Numeric}

\item{Divergence}{Numeric}

\item{Rho}{cylinder radius - Untested. Used if Cylinder is missing.}

\item{Jugacy}{untested}

\item{L}{Display height of cylinder. Used if Cylinder is missing.}

\item{A}{Cylinder object.}
}
\description{
Create a new PhyllotaxisGenetic object
}
