% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\docType{class}
\name{PhyllotaxisGenetic-class}
\alias{PhyllotaxisGenetic-class}
\title{A class representing lattices defined by their divergence and rise}
\description{
A class representing lattices defined by their divergence and rise
}
\section{Slots}{

\describe{
\item{\code{Rise}}{Numeric}

\item{\code{Divergence}}{Numeric}

\item{\code{Jugacy}}{Numeric (untested)}

\item{\code{Cylinder}}{object of class Cylinder}
}}

