% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneratingIntervals.R
\name{plotGeneratingIntervals}
\alias{plotGeneratingIntervals}
\title{Plot the regions on which d is generating/opposed for a range of m,n pairs}
\usage{
plotGeneratingIntervals(mnlist)
}
\arguments{
\item{mnlist}{A list of integer (m,n) pairs}
}
\description{
Plot the regions on which d is generating/opposed for a range of m,n pairs
}
