% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\name{plotParastichyPairs}
\alias{plotParastichyPairs}
\title{Plot a lattice and a parastichy pair}
\usage{
plotParastichyPairs(PG, y, m, n, doParallelogram = 0, dodash = TRUE)
}
\arguments{
\item{PG}{A lattice object}

\item{m}{Order of the parastichy to plot}

\item{n}{Order of the parastichy to plot}

\item{doParallelogram}{UNUSED}

\item{dodash}{UNUSED}
}
\description{
Plot a lattice and a parastichy pair
}
