% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\name{plotParastichySegment}
\alias{plotParastichySegment}
\title{Plot a segement of a parastichy line without wrapping}
\usage{
plotParastichySegment(PG, parastichyCount, through, length, gp)
}
\arguments{
\item{PG}{A lattice object}

\item{parastichyCount}{The order of the parastichy}

\item{through}{The sequence number of the point to pass through}

\item{length}{The length of the parastichy line (in cylinder units, ugh)}

\item{gp}{A gpar object}
}
\description{
Plot a segement of a parastichy line without wrapping
}
