% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricPhyllotaxis.R
\name{plotPhyllotaxis}
\alias{plotPhyllotaxis}
\title{Plot a phyllotactic lattice}
\usage{
plotPhyllotaxis(PG, plotPrincipals = FALSE, ...)
}
\arguments{
\item{PG}{A lattice object which can be coerced to class PhyllotaxisGenetic}

\item{plotPrincipals}{Default FALSE. If TRUE, plot the principal parastichy vectors.}

\item{doRepeat}{=0,}

\item{doNumbers}{=FALSE,}

\item{pointSize}{=0.2,}

\item{doCircles}{=FALSE,}

\item{doFacets}{=FALSE}
}
\description{
Plot a phyllotactic lattice
}
