% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvanIterson.R
\name{plotvanIterson}
\alias{plotvanIterson}
\title{Plot elements of lattice space including the van Iterson tree}
\usage{
plotvanIterson(vpairs, labellines = FALSE, labelregions = TRUE, vpairlabels,
  xlim, ylim, onlyPrimary = FALSE, doCircles = TRUE,
  primaryColour = "black", doTriplePoints = FALSE)
}
\arguments{
\item{vpairs}{A data frame in which the first two columns are pair of parastichy numbers. If the third column is FALSE, do not label the pair}

\item{labellines}{Label the branches of the van Iterson tree. Default FALSE}

\item{labelregions}{Label the regions by their principal parastichy values}

\item{vpairlabels}{A data frame in which the first two columns are the parastichy numbers for labelling the region and the second two the (d,h) coordinates where the labels sohuld be placed}

\item{xlim, ylim}{A pair of reals defining the plotting area to use}

\item{onlyPrimary}{If true, do not plot lines on which a pair is van Iterson but not primary}

\item{doCircles}{If true, plot entire circles on which van Iterson branches lie}

\item{primaryColour}{The colour for the primary branch}

\item{doTriplePoints}{If true, plot a marker at each triplepoint}
}
\description{
Plot elements of lattice space including the van Iterson tree
}
