% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{vanItersonDSquared}
\alias{vanItersonDSquared}
\title{The square of the disk radius D for a van Iterson diagram}
\usage{
vanItersonDSquared(m, n, d)
}
\arguments{
\item{m, n}{Integer. Orders of a generating parastichy pair}

\item{d}{Divergence of the lattice. 0<d<1.}
}
\description{
The square of the disk radius D for a van Iterson diagram
}
