% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{vanItersonhSquared}
\alias{vanItersonhSquared}
\title{The square of the rise h for a van Iterson diagram}
\usage{
vanItersonhSquared(m, n, d)
}
\arguments{
\item{m, n}{Integer. Orders of a generating parastichy pair}

\item{d}{Divergence of the lattice. 0<d<1.}
}
\description{
The square of the rise h for a van Iterson diagram
}
