% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{vanItersonTriplePoint}
\alias{vanItersonTriplePoint}
\title{The (h,d) coordinates of the (m,n,m+n) triple point in a van Iterson diagram}
\usage{
vanItersonTriplePoint(m, n)
}
\arguments{
\item{m, n}{Integers}
}
\description{
The (h,d) coordinates of the (m,n,m+n) triple point in a van Iterson diagram
}
