% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{vanItersonTriplePointL}
\alias{vanItersonTriplePointL}
\title{The van Iterson triple point for m,n with smaller 0<d<1/2}
\usage{
vanItersonTriplePointL(m, n)
}
\arguments{
\item{m, n}{Integer}
}
\description{
The van Iterson triple point for m,n with smaller 0<d<1/2
}
