% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanIterson.R
\name{vanItersonTriplePointR}
\alias{vanItersonTriplePointR}
\title{The van Iterson triple point for m,n with larger 0<d<1/2}
\usage{
vanItersonTriplePointR(m, n)
}
\arguments{
\item{m, n}{Integer}
}
\description{
The van Iterson triple point for m,n with larger 0<d<1/2
}
