% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyllotaxisGeneratingIntervals.R
\name{windingNumbers}
\alias{windingNumbers}
\title{Calculate winding numbers.}
\usage{
windingNumbers(m, n, Delta = 1)
}
\arguments{
\item{m, n}{Positive integer}

\item{Delta}{Plus or minus 1.}
}
\description{
The winding numbers (u,v) for a pair (m,n) satisfy mv-nu=Delta for Delta=+/- 1, and if 1<m,n they are in the range
0<u<m, 0<v<n. windingNumbers returns 0,1 when called with 1,n and m-1,0 when called with m,1.
Winding numbers can only be found if m and n are coprime and the function throws an error if not.
If
}
